rooms for server

+++ src/Util.cc
@@ -2,7 +2,7 @@
 
 std::string Util::subStr(const std::string& str, int n)
 {
-	if(str.length() < n)
+	if((int)str.length() < n)
 		return str;
 	return str.substr(0, n);
 }
